pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "irvi"
version = "0.1.0"
description = "Multimodal Transformer-based VAE for gene expression and TCR data"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Wendao Liu", email = "liuwd15@gmail.com"},
]
maintainers = [
    {name = "Wendao Liu", email = "liuwd15@gmail.com"},
]
keywords = [
    "single-cell",
    "genomics",
    "t-cell-receptor",
    "variational-autoencoder",
    "transformer",
    "immune-repertoire",
    "bioinformatics",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "scvi-tools>=1.0.0",
    "torch>=1.12.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "anndata>=0.8.0",
    "scanpy>=1.8.0",
    "rich>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "flake8>=4.0",
    "isort>=5.0",
    "pre-commit>=2.0",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=0.17",
    "nbsphinx>=0.8",
    "sphinx-autodoc-typehints>=1.12",
]
examples = [
    "jupyter>=1.0",
    "matplotlib>=3.3",
    "seaborn>=0.11",
    "plotly>=5.0",
]

[project.urls]
Homepage = "https://github.com/liuwd15/irvi"
Documentation = "https://irvi.readthedocs.io/"
Repository = "https://github.com/liuwd15/irvi.git"
Issues = "https://github.com/liuwd15/irvi/issues"
Changelog = "https://github.com/liuwd15/irvi/blob/main/CHANGELOG.md"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["."]
include = ["irvi*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
irvi = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=irvi",
    "--cov-report=term-missing",
]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["irvi"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]